+++ src/test/elements/box/assets/UpdateCustomFields.json
@@ -1,5 +1,5 @@
 {
-  "path": "/" ,
+  "path": "/",
   "value": "madhuri",
   "scope": "enterprise"
 }
