+++ src/test/elements/box/assets/customFields.json
@@ -1,4 +1,4 @@
 {
 "scope":"enterprise",
-"template":"customer"
+"template":""
 }
