+++ src/test/elements/box/files.js
@@ -6,9 +6,11 @@ const cloud = require('core/cloud');
 const faker = require('faker');
 const tools = require('core/tools');
 const payload = tools.requirePayload(`${__dirname}/assets/customFields.json`);
+const temPayload = tools.requirePayload(`${__dirname}/assets/template.json`);
 const templateKeyPayload = tools.requirePayload(`${__dirname}/assets/UpdateCustomFields.json`);
 const commentPayload1 = tools.requirePayload(`${__dirname}/assets/comment.json`);
 const commentPayload2 = tools.requirePayload(`${__dirname}/assets/comment.json`);
+const updatePayload = tools.requirePayload(`${__dirname}/assets/PutCustomFields.json`);
 const folderPayload = require('./assets/folders');
 
 const randomInt = tools.randomInt();
@@ -19,34 +21,77 @@ const lock = {
   "is_download_prevented": false,
   "expires_at": "2030-12-12T10:55:30-08:00"
 };
-const updatePayload =tools.requirePayload(`${__dirname}/assets/PutCustomFields.json`);
+
 suite.forElement('documents', 'files', (test) => {
-  let id, path, fileTagPayload, fileUpdateTagPayload, templateKey;
+  let id, path, name, fileTagPayload, fileUpdateTagPayload, templateKey;
+  let file = `${__dirname}/assets/brady.jpg`;
 
   before(done => {
-    return cloud.withOptions({ qs: { path: `/brady-${faker.address.zipCode()}.jpg` } }).postFile(test.api, `${__dirname}/../assets/brady.jpg`)
+    return cloud.withOptions({ qs: { path: `/brady-${faker.address.zipCode()}.jpg` } }).postFile(test.api, `${__dirname}/assets/brady.jpg`)
       .then(r => {
         id = r.body.id;
         path = r.body.path;
+        name = r.body.name;
         fileTagPayload = r.body;
         fileTagPayload.tags = ["fileTag1", "fileTag2"];
       })
-      .then(r =>  cloud.withOptions({ qs: { scope: "enterprise" } }).get('/hubs/documents/custom-fields/templates'))
+      .then(r =>  cloud.post(`/custom-fields-templates`, temPayload))
           .then(r => {
-            templateKey = r.body[0].templateKey;
-            payload.template = r.body[0].templateKey;
-            updatePayload.template= r.body[0].templateKey;
-            templateKeyPayload.path="/"+r.body[0].fields[0].key;
+        templateKey = r.body.templateKey;
+        payload.template = r.body.templateKey;
+        updatePayload.template= r.body.templateKey;
+        updatePayload.path = "/" + r.body.fields[0].key;
+        templateKeyPayload.path= "/" + r.body.fields[0].key;
           })
           .then(r => done());
-  }
-);
-  after(() => cloud.delete(`${test.api}/${id}`));
+  });
+  
+  after(() => cloud.delete(`${test.api}/${id}`)
+    .then(() => cloud.withOptions({ qs: {scope: "enterprise"}}).delete(`/custom-fields-templates/${templateKey}`))
+  );
+  
   afterEach(done => {
-    //We were getting a 429 before this
+    // To avoid 429 rate limiting
     setTimeout(done, 2500);
   });
 
+
+  it('should allow CRD /files and RD /files/:id', () => {
+    let pathFile,
+        query = { path: `/brady-${tools.randomStr('abcdefghijklmnopqrstuvwxyz1234567890', 10)}.jpg` };
+    return cloud.get(`/hubs/documents/files/${id}`)
+      .then(() => cloud.withOptions({ qs: Object.assign({}, { path: `${query.path}-ayy` }) }).postFile('/hubs/documents/files', file))
+      .then(r => pathFile = r.body)
+      .then(() => cloud.withOptions({ qs: { path: pathFile.path } }).get('/hubs/documents/files'))
+      .then(r => cloud.withOptions({ qs: { path: pathFile.path } }).delete('/hubs/documents/files'));
+  });
+
+  it('should allow GET /files/links and /files/:id/links', () => {
+    return cloud.get(`/hubs/documents/files/${id}/links`)
+      .then(() => cloud.withOptions({ qs: { path: path } }).get('/hubs/documents/files/links'));
+  });
+
+  it('should allow RU /files/metadata and RU /files/:id/metadata', () => {
+    let updatedFile,
+      fileTemp = { path: `/a-${name}` };
+    return cloud.withOptions({ qs: { path: path } }).get('/hubs/documents/files/metadata')
+      .then(r => cloud.withOptions({ qs: { path: path } }).patch('/hubs/documents/files/metadata', fileTemp))
+      .then(r => updatedFile = r.body)
+      .then(r => cloud.patch(`/hubs/documents/files/${updatedFile.id}/metadata`, fileTagPayload))
+      .then(r => cloud.get(`/hubs/documents/files/${id}/metadata`));
+  });
+
+  it('should allow POST /files/copy and POST /files/:id/copy', () => {
+    const copy1 = { path: '/churrosCopy1' + tools.random() },
+      copy2 = { path: '/churrosCopy2' + tools.random() };
+    let fileCopy1, fileCopy2;
+    return cloud.withOptions({ qs: { path: path } }).post('/hubs/documents/files/copy', copy1)
+      .then(r => fileCopy1 = r.body)
+      .then(() => cloud.withOptions({ qs: { path: path } }).post('/hubs/documents/files/copy', copy2))
+      .then(r => fileCopy2 = r.body)
+      .then(() => cloud.delete(`/hubs/documents/files/${fileCopy1.id}`))
+      .then(() => cloud.delete(`/hubs/documents/files/${fileCopy2.id}`));
+  });
   it('should allow POST /files by providing folder id', () => {
     let folderId, fileId;
     return cloud.post('/folders', folderPayload)
@@ -103,22 +148,17 @@ suite.forElement('documents', 'files', (test) => {
 
 
   it('should allow CRUDS for /files/:id/custom-fields', () => {
-
       return cloud.get(`/hubs/documents/files/${id}/custom-fields`)
       .then(r => cloud.post(`/hubs/documents/files/${id}/custom-fields`, payload))
       .then(r => cloud.put(`/hubs/documents/files/${id}/custom-fields`, updatePayload))
       .then(r => cloud.patch(`/hubs/documents/files/${id}/custom-fields`, updatePayload))
       .then(r => cloud.withOptions({ qs: { scope: "enterprise" } }).get(`/hubs/documents/files/${id}/custom-fields/${templateKey}`))
-      .then(r => cloud.patch(`/hubs/documents/files/${id}/custom-fields/${templateKey}`, templateKeyPayload))
-      .then(r => cloud.withOptions({ qs: { scope: "enterprise" } }).delete(`/hubs/documents/files/${id}/custom-fields/${templateKey}`));
+      .then(r => cloud.patch(`/hubs/documents/files/${id}/custom-fields/${templateKey}`, templateKeyPayload));
   });
 
   it('should allow RUD for /files/{id}/custom-fields-templates/{templateKeyId}/custom-fields', () => {
-
-      return cloud.post(`/hubs/documents/files/${id}/custom-fields`, payload)
-      .then(r => cloud.withOptions({ qs: { scope: "enterprise" } }).get(`/hubs/documents/files/${id}/custom-fields-templates/${templateKey}/custom-fields`))
-      .then(r => cloud.patch(`/hubs/documents/files/${id}/custom-fields-templates/${templateKey}/custom-fields`, templateKeyPayload))
-      .then(r => cloud.withOptions({ qs: { scope: "enterprise" } }).delete(`/hubs/documents/files/${id}/custom-fields-templates/${templateKey}/custom-fields`));
+      return cloud.withOptions({ qs: { scope: "enterprise" } }).get(`/hubs/documents/files/${id}/custom-fields-templates/${templateKey}/custom-fields`)
+      .then(r => cloud.patch(`/hubs/documents/files/${id}/custom-fields-templates/${templateKey}/custom-fields`, templateKeyPayload));
   });
 
 
@@ -127,13 +167,10 @@ suite.forElement('documents', 'files', (test) => {
    * While we don't offer the POST /revisions endpoint we'll need to hardcode the file data
    */
   it('should allow RS for documents/files/:id/revisions', () => {
-    const fileId = 288284426512;
+    const fileId = 310815318732;
     let revisionId;
     return cloud.get(`${test.api}/${fileId}/revisions`)
-      .then(r => {
-        expect(r.body[0]).to.contain.key('id');
-        revisionId = r.body[0].id;
-      })
+      .then(r => revisionId = r.body[0].id)
       .then(() => cloud.get(`${test.api}/${fileId}/revisions/${revisionId}`));
   });
 
@@ -141,10 +178,7 @@ suite.forElement('documents', 'files', (test) => {
     let options = { qs: { path: '/ChurrosRevisionsAPIDontDelete/revisionstestfile.txt' } };
     let revisionId;
     return cloud.withOptions(options).get(`${test.api}/revisions`)
-      .then(r => {
-        expect(r.body[0]).to.contain.key('id');
-        revisionId = r.body[0].id;
-      })
+      .then(r => revisionId = r.body[0].id)
       .then(() => cloud.withOptions(options).get(`${test.api}/revisions/${revisionId}`));
   });
